Make it possible to add a new section to default blog in the theme editor.
In this lesson we're going to migrate the blog template from liquid to JSON template, so go back to your customizer page and open the dropdown with
the blog template, > blogs > default blog. In the left sidebar, you can see that it's not yet modular, so we can't customize or add sections.

That's what we'll update in this lesson.

In vs code, templates/blog.liquid, we'll move (cut-paste) all the code from this file to a new section, so:
- add a new file to the sections folder, name it template-blog.liquid and move the code
- delete blog.liquid file
- now make the json template file, back in /templates folder, touch blog.json
- inside blog.json, add an object: {}
- add the "sections" sub-object and "order" string:
    "sections": {

    },
    "order": ""
- then we'll add "template" inside "sections", then copy it and apply it to "order", as the first member of the array:
    {
      "sections": {
          "template": { <---- "template" here is referred to as "section id"

          }
      },
      "order": [
          "template"
      ]
    }
- now we'll add the "type", and reference the template-blog.liquid file:
    {
      "sections": {
          "template": {
              "type": "template-blog"
          }
      },
      "order": [
          "template"
      ]
    }

That is it.